# Make $nu constant