--- a/crates/nu-cli/tests/support/completions_helpers.rs
+++ b/crates/nu-cli/tests/support/completions_helpers.rs
@@ -1,10 +1,10 @@
 use std::path::PathBuf;
 
-use nu_engine::eval_block;
+use nu_engine::{eval_block, eval_nu_variable};
 use nu_parser::parse;
 use nu_protocol::{
     engine::{EngineState, Stack, StateWorkingSet},
-    PipelineData, ShellError, Span, Value,
+    PipelineData, ShellError, Span, Value, NU_VARIABLE_ID,
 };
 use nu_test_support::fs;
 use reedline::Suggestion;
@@ -27,6 +27,10 @@ pub fn new_engine() -> (PathBuf, String, EngineState, Stack) {
 
     // Create a new engine with default context
     let mut engine_state = create_default_context();
+
+    // Add $nu
+    let nu_const = eval_nu_variable(&engine_state, Span::test_data()).expect("Failed creating $nu");
+    engine_state.set_variable_const_val(NU_VARIABLE_ID, nu_const);
 
     // New stack
     let mut stack = Stack::new();