--- a/crates/nu-cli/tests/support/completions_helpers.rs
+++ b/crates/nu-cli/tests/support/completions_helpers.rs
@@ -1,9 +1,10 @@
 use std::path::PathBuf;
 
-use nu_engine::{eval_block, eval_nu_variable};
+use nu_engine::eval_block;
 use nu_parser::parse;
 use nu_protocol::{
     engine::{EngineState, Stack, StateWorkingSet},
+    eval_const::create_nu_constant,
     PipelineData, ShellError, Span, Value, NU_VARIABLE_ID,
 };
 use nu_test_support::fs;
@@ -29,7 +30,8 @@ pub fn new_engine() -> (PathBuf, String, EngineState, Stack) {
     let mut engine_state = create_default_context();
 
     // Add $nu
-    let nu_const = eval_nu_variable(&engine_state, Span::test_data()).expect("Failed creating $nu");
+    let nu_const =
+        create_nu_constant(&engine_state, Span::test_data()).expect("Failed creating $nu");
     engine_state.set_variable_const_val(NU_VARIABLE_ID, nu_const);
 
     // New stack