--- a/crates/nu-system/Cargo.toml
+++ b/crates/nu-system/Cargo.toml
@@ -15,6 +15,7 @@ bench = false
 [dependencies]
 libc = "0.2"
 log = "0.4"
+sysinfo = "0.29"
 
 [target.'cfg(target_family = "unix")'.dependencies]
 nix = { version = "0.26", default-features = false, features = ["fs", "term", "process", "signal"] }
@@ -32,31 +33,31 @@ ntapi = "0.4"
 once_cell = "1.18"
 winapi = { version = "0.3", features = [
     "tlhelp32",
-    "fileapi",
-    "handleapi",
-    "ifdef",
-    "ioapiset",
-    "minwindef",
-    "pdh",
-    "psapi",
-    "synchapi",
-    "sysinfoapi",
-    "winbase",
-    "winerror",
-    "winioctl",
-    "winnt",
-    "oleauto",
-    "wbemcli",
-    "rpcdce",
-    "combaseapi",
-    "objidl",
-    "powerbase",
-    "netioapi",
-    "lmcons",
-    "lmaccess",
-    "lmapibuf",
-    "memoryapi",
-    "shellapi",
-    "std",
-    "securitybaseapi",
-] }
+        "fileapi",
+        "handleapi",
+        "ifdef",
+        "ioapiset",
+        "minwindef",
+        "pdh",
+        "psapi",
+        "synchapi",
+        "sysinfoapi",
+        "winbase",
+        "winerror",
+        "winioctl",
+        "winnt",
+        "oleauto",
+        "wbemcli",
+        "rpcdce",
+        "combaseapi",
+        "objidl",
+        "powerbase",
+        "netioapi",
+        "lmcons",
+        "lmaccess",
+        "lmapibuf",
+        "memoryapi",
+        "shellapi",
+        "std",
+        "securitybaseapi",
+        ] }