--- a/crates/nu-system/src/lib.rs
+++ b/crates/nu-system/src/lib.rs
@@ -3,6 +3,7 @@ mod foreground;
 mod linux;
 #[cfg(target_os = "macos")]
 mod macos;
+pub mod os_info;
 #[cfg(target_os = "windows")]
 mod windows;