<!--
if this PR closes one or more issues, you can automatically link the PR with
them by using one of the [*linking keywords*](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue#linking-a-pull-request-to-an-issue-using-a-keyword), e.g.
- this PR should close #xxxx
- fixes #xxxx

you can also mention related issues, PRs or discussions!
-->

# Description
<!--
Thank you for improving Nushell. Please, check our [contributing guide](../CONTRIBUTING.md) and talk to the core team before making major changes.

Description of your pull request goes here. **Provide examples and/or screenshots** if your changes affect the user experience.
-->

This PR makes `$nu` a parse-time constant. The benefits are two-fold:

1. `$nu` is now usable in more situations. For example, reloading the config file is now more convenient than ever:
```nushell
source $nu.config-path
```
2. Performance. Evaluating `$nu` is now approximately 15x faster:
```nushell
> use std
> std bench { 1..1000 | each { $nu } } --pretty
# before:
80ms 681µs 622ns +/- 1ms 185µs 491ns
# after:
5ms 211µs 764ns +/- 1ms 39µs 16ns
```

The only remaining question is where to move the `eval_nu_variable()` since it is now not a part of the engine. I'm leaning towards `nu-protocol` which would move the `sysinfo` dependency from `nu-engine` to `nu-protocol`, alongside to introducing `nu-path` dependency.

# User-Facing Changes
<!-- List of all changes that impact the user experience here. This helps us keep track of breaking changes. -->

Enhances `$nu` usability, but everything should work exactly as before.

# Tests + Formatting
<!--
Don't forget to add tests that cover your changes.

Make sure you've run and fixed any issues with these commands:

- `cargo fmt --all -- --check` to check standard code formatting (`cargo fmt --all` applies these changes)
- `cargo clippy --workspace -- -D warnings -D clippy::unwrap_used` to check that you're using the standard code style
- `cargo test --workspace` to check that all tests pass (on Windows make sure to [enable developer mode](https://learn.microsoft.com/en-us/windows/apps/get-started/developer-mode-features-and-debugging))
- `cargo run -- -c "use std testing; testing run-tests --path crates/nu-std"` to run the tests for the standard library

> **Note**
> from `nushell` you can also use the `toolkit` as follows
> ```bash
> use toolkit.nu  # or use an `env_change` hook to activate it automatically
> toolkit check pr
> ```
-->

# After Submitting
<!-- If your PR had any user-facing changes, update [the documentation](https://github.com/nushell/nushell.github.io) after the PR is merged, if necessary. This will help us keep the docs up to date. -->